Added new columns to at_bat table
Added following colums to the at_bat table:
- ab_result : string(32)
- traj : string(8)
- fielder : string(8)
- spray_x : Float
- spray_y : Float
- hard_hit : Boolean
- inning : Integer
- notes : String(128)

All of these columns came from the pitches table. They are only
provided a value once per at bat, so they are better served being in the
at_bat table.

The biggest problem with moving over the columns is that nearly
every function and class in our system will have to be updated, because
most of the functions depend on the presence of these fields in the
pitches table... therefore I will not be deleting these columns from
the pitches table in the near future.