Clean up merge of update-database-schema
I had to do some flask migrate magic, because for some reason at
this point, there were 3 migrations heads to the databse. I added
a merge point for the migration heads which can be seen if the
command "flask db history" is run.

Currently, when running "flask db upgrade", there is an error
where flask migrate attempts to add duplicates of the
current_player and current_coach columns in the database. It does
not overite data, just produces a bunch of error text. It also puts
the other changes into place as if it were normal... I wil figure
how to fix it tomorrow.