Update database in prep for sprays
Cleaned up the database history errors so that all of the past
heads were merged.

Created 2 new columns in the opponents table: mascot and ncaa_id.
Mascot is just meant to hold the mascot that we already have in the
name, and ncaa_id holds the id that ncaa.com assigns to a team that
brian scraped.

Created 1 new column in season, which will hold the ncaa year id
scraped by brian for each playing season, and null for fall and
summer seasons.